# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: norms
    description: Comparison with other participants (Remembrance-internal norms)
  - name: users
    description: POC users (no auth; identified by email)
  - name: scores
    description: Saved score snapshots for a user
paths:
  /waitlist:
    post:
      operationId: joinWaitlist
      summary: Join the Remembrance early-access waitlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WaitlistInput"
      responses:
        "200":
          description: Signup saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistResult"
        "400":
          description: Invalid email
        "429":
          description: Please wait before retrying
        "503":
          description: Temporarily unavailable
  /norms/compare:
    post:
      operationId: compareWithNorms
      tags: [norms]
      summary: Compare a result with other participants of the same age and sex
      description: >
        Stores the participant's first completed administration of a test as a
        de-identified reference result (later ones are ignored), then compares the
        submitted score fields with the participant's age/sex group. Groups
        smaller than minCohort are reported as insufficient.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NormSubmission"
      responses:
        "200":
          description: Comparison (and whether this result was added to the reference)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NormComparison"
        "400":
          description: Invalid submission
        "429":
          description: Please wait before retrying
        "503":
          description: Comparison unavailable (no database configured)
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user, or return the existing one for that email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserInput"
      responses:
        "200":
          description: User created or already existed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResult"
        "400":
          description: Invalid input
        "503":
          description: Temporarily unavailable
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResult"
        "404":
          description: No user with that id
  /users/{id}/scores:
    post:
      operationId: saveScore
      tags: [scores]
      summary: Save a score snapshot for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveScoreInput"
      responses:
        "200":
          description: Score saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreResult"
        "400":
          description: Invalid input
        "404":
          description: No user with that id
    get:
      operationId: listScores
      tags: [scores]
      summary: List a user's saved scores, newest first
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Score history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoreResult"
        "404":
          description: No user with that id
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
components:
  schemas:
    NormSubmission:
      type: object
      required: [participantId, testId, formId, specVersion, scorerVersion, equated, channel, administrationNumber, completed, birthYear, sex, fields]
      properties:
        participantId:
          type: string
          minLength: 1
          maxLength: 128
        testId:
          type: string
          maxLength: 64
        formId:
          type: string
          maxLength: 64
        specVersion:
          type: string
          maxLength: 32
        scorerVersion:
          type: string
          maxLength: 32
        equated:
          type: boolean
        channel:
          type: string
          enum: [web, phone]
        administrationNumber:
          type: integer
          minimum: 1
        completed:
          type: boolean
        birthYear:
          type: integer
          minimum: 1900
          maximum: 2100
        sex:
          type: string
          enum: [female, male, undisclosed]
        educationYears:
          type: integer
          minimum: 0
          maximum: 30
        fields:
          type: object
          additionalProperties:
            type: [number, "null"]
    NormComparison:
      type: object
      required: [stored, cohort, minCohort, results]
      properties:
        stored:
          type: boolean
          description: True when this result was added to the reference (first completed administration).
        cohort:
          type: string
          description: Plain-language group, e.g. "women aged 60–69".
        minCohort:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/NormFieldComparison"
    NormFieldComparison:
      type: object
      required: [field, label, value, n, percentile, median]
      properties:
        field:
          type: string
        label:
          type: string
        value:
          type: number
        n:
          type: integer
          description: People in the group with this score.
        percentile:
          type: [integer, "null"]
          description: Percent of the group this result is better than (lower times are better; ties count half); null when the group is too small.
        median:
          type: [number, "null"]
    WaitlistInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
    WaitlistResult:
      type: object
      required: [message]
      properties:
        message:
          type: string
    CreateUserInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        name:
          type: string
          maxLength: 128
    UserResult:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
    SaveScoreInput:
      type: object
      required: [testId, fields]
      properties:
        testId:
          type: string
          maxLength: 64
        composite:
          type: [integer, "null"]
          minimum: 0
          maximum: 100
        fields:
          type: object
          additionalProperties:
            type: [number, "null"]
    ScoreResult:
      type: object
      required: [id, userId, testId, fields, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        testId:
          type: string
        composite:
          type: [integer, "null"]
        fields:
          type: object
          additionalProperties:
            type: [number, "null"]
        createdAt:
          type: string
          format: date-time
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status