# lib/api-zod/src/generated/types/createUserInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateUserInput {
  /** @maxLength 254 */
  email: string;
  /** @maxLength 128 */
  name?: string;
}