# lib/api-zod/src/generated/types/saveScoreInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SaveScoreInputFields } from './saveScoreInputFields';

export interface SaveScoreInput {
  /** @maxLength 64 */
  testId: string;
  /**
     * @minimum 0
     * @maximum 100
     * @nullable
     */
  composite?: number | null;
  fields: SaveScoreInputFields;
}