# lib/api-zod/src/generated/types/scoreResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScoreResultFields } from './scoreResultFields';

export interface ScoreResult {
  id: number;
  userId: number;
  testId: string;
  /** @nullable */
  composite?: number | null;
  fields: ScoreResultFields;
  createdAt: Date;
}