# lib/api-zod/src/generated/types/userResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserResult {
  id: number;
  email: string;
  /** @nullable */
  name?: string | null;
  createdAt: Date;
}